cbt-rag-service/pyproject.toml
[project]
name = "cbt-rag-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "chromadb>=1.5.7",
    "fastapi>=0.135.3",
    "google-genai>=1.72.0",
    "httpx>=0.28.1",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-text-splitters>=1.1.1",
    "pydantic>=2.12.5",
    "pypdf>=6.10.0",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.4.0",
    "uvicorn>=0.44.0",
]
